Correct typo in example
The typo is a rather frustrating one. I had overlooked that it should say `output` not `ouput`. The result is that none of my settings were having any effect. After stepping through the code with the debugger I realised that my config was wrong and I had copied and pasted it from `README.md`.